requirements.txt
torch
numpy
faiss-gpu
cv2
torchvision
streamlit
open-clip-torch
gradio
